refactor: Delete NonceCacheService and combine it with VCICacheService
