+++ certify-service/src/main/java/io/mosip/certify/services/NonceServiceImpl.java
@@ -2,48 +2,42 @@ package io.mosip.certify.services;
 
 import io.mosip.certify.core.dto.NonceResponse;
 import io.mosip.certify.core.dto.VCIssuanceTransaction;
-import io.mosip.certify.core.exception.CertifyException;
 import io.mosip.certify.core.spi.NonceService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 
 @Service
 public class NonceServiceImpl implements NonceService {
 
-    private NonceCacheService nonceCacheService;
+    private VCICacheService vciCacheService;
 
     @Value("${mosip.certify.cnonce-expire-seconds:300}")
     private int cNonceExpiresInSeconds;
 
-    public NonceServiceImpl(
-                            NonceCacheService nonceCacheService) {
-        this.nonceCacheService = nonceCacheService;
+    public NonceServiceImpl(VCICacheService vciCacheService) {
+        this.vciCacheService = vciCacheService;
     }
 
     @Override
     public NonceResponse generateNonce() {
         String cNonce = generateCNonce();
-        VCIssuanceTransaction transaction = createNonceTransaction(cNonce);
+        createNonceTransaction(cNonce);
         return new NonceResponse(cNonce);
     }
 
     public String generateCNonce() {
-        String cNonce = java.util.UUID.randomUUID().toString();
-        return cNonce;
+        return java.util.UUID.randomUUID().toString();
     }
 
-    private VCIssuanceTransaction createNonceTransaction(String cNonce) {
-        try {
-            Long cNonceIssuedTime = LocalDateTime.now(ZoneOffset.UTC).toEpochSecond(ZoneOffset.UTC);
-            VCIssuanceTransaction transaction = new VCIssuanceTransaction();
-            transaction.setCNonce(cNonce);
-            transaction.setCNonceIssuedEpoch(cNonceIssuedTime);
-            transaction.setCNonceExpireSeconds(cNonceExpiresInSeconds);
-            return nonceCacheService.setNonceTransaction(cNonce, transaction);
-        } catch (CertifyException e) {
-            throw e;
-        }
+    private void createNonceTransaction(String cNonce) {
+        long cNonceIssuedTime = LocalDateTime.now(ZoneOffset.UTC).toEpochSecond(ZoneOffset.UTC);
+        VCIssuanceTransaction transaction = new VCIssuanceTransaction();
+        transaction.setCNonce(cNonce);
+        transaction.setCNonceIssuedEpoch(cNonceIssuedTime);
+        transaction.setCNonceExpireSeconds(cNonceExpiresInSeconds);
+        vciCacheService.setNonceTransaction(cNonce, transaction);
     }
 }
