+++ certify-service/src/main/java/io/mosip/certify/services/PreAuthIssuanceServiceImpl.java
@@ -35,6 +35,10 @@ public class PreAuthIssuanceServiceImpl implements DataProviderPlugin {
         PreAuthTransaction cachedTransaction = vciCacheService.getPreAuthTransaction(accessTokenHash);
 
         // Validate claims in the cached transaction
+        if (cachedTransaction == null) {
+            log.error("No cached transaction found for access token hash: {}", accessTokenHash);
+            throw new DataProviderExchangeException("Cached transaction is missing");
+        }
         if (cachedTransaction.getClaims() == null || cachedTransaction.getClaims().isEmpty()) {
             log.error("No claims found in cached transaction for access token hash: {}", accessTokenHash);
             throw new DataProviderExchangeException("Cached transaction claims are null or empty");
