+++ certify-service/src/main/resources/application-local.properties
@@ -151,7 +151,7 @@ mosip.certify.cache.security.algorithm-name=AES/ECB/PKCS5Padding
 #spring.data.redis.password=redis
 
 spring.cache.type=simple
-mosip.certify.cache.names=userinfo,vcissuance,templatecache,certificatedatacache,credentialConfig,renderTemplate,jwks,preAuthCodeCache,credentialOfferCache,issuerMetadataCache,nonce
+mosip.certify.cache.names=userinfo,vcissuance,templatecache,certificatedatacache,credentialConfig,renderTemplate,jwks,preAuthCodeCache,credentialOfferCache,issuerMetadataCache,nonce,preAuthCacheTxn
 spring.cache.cache-names=${mosip.certify.cache.names}
 management.health.redis.enabled=false
 
@@ -170,10 +170,10 @@ mosip.certify.common.cache-expire-seconds=3600
 mosip.certify.jwks-cache-expire-seconds=300
 # Cache size setup is applicable only for 'simple' cache type.
 # Cache size configuration will not be considered with 'Redis' cache type
-mosip.certify.cache.size={'userinfo': 200, 'vcissuance' : 2000, 'templatecache': 20, 'jwks': 1, 'preAuthCodeCache': 1000, 'credentialOfferCache': 1000, 'issuerMetadataCache': 10, 'nonce': 1000}
+mosip.certify.cache.size={'userinfo': 200, 'vcissuance' : 2000, 'templatecache': 20, 'jwks': 1, 'preAuthCodeCache': 1000, 'credentialOfferCache': 1000, 'issuerMetadataCache': 10, 'nonce': 1000, 'preAuthCacheTxn' : 2000}
 
 # Cache expire in seconds is applicable for both 'simple' and 'Redis' cache type
-mosip.certify.cache.expire-in-seconds={'userinfo': ${mosip.certify.access-token-expire-seconds}, 'vcissuance': ${mosip.certify.access-token-expire-seconds}, 'templatecache': ${mosip.certify.templatecache-expire-seconds}, 'certificatedatacache': ${mosip.certify.certificatedatacache-expire-seconds}, 'credentialConfig': ${mosip.certify.common.cache-expire-seconds}, 'renderTemplate': ${mosip.certify.common.cache-expire-seconds}, 'jwks': ${mosip.certify.jwks-cache-expire-seconds}, 'preAuthCodeCache': ${mosip.certify.pre-auth.cache-expire-seconds}, 'credentialOfferCache': ${mosip.certify.credential-offer.cache-expire-seconds}, 'issuerMetadataCache': ${mosip.certify.metadata.cache-expire-seconds}, 'nonce': ${mosip.certify.cnonce.cache-expire-seconds}}
+mosip.certify.cache.expire-in-seconds={'userinfo': ${mosip.certify.access-token-expire-seconds}, 'vcissuance': ${mosip.certify.access-token-expire-seconds}, 'templatecache': ${mosip.certify.templatecache-expire-seconds}, 'certificatedatacache': ${mosip.certify.certificatedatacache-expire-seconds}, 'credentialConfig': ${mosip.certify.common.cache-expire-seconds}, 'renderTemplate': ${mosip.certify.common.cache-expire-seconds}, 'jwks': ${mosip.certify.jwks-cache-expire-seconds}, 'preAuthCodeCache': ${mosip.certify.pre-auth.cache-expire-seconds}, 'credentialOfferCache': ${mosip.certify.credential-offer.cache-expire-seconds}, 'issuerMetadataCache': ${mosip.certify.metadata.cache-expire-seconds}, 'nonce': ${mosip.certify.cnonce.cache-expire-seconds}, 'preAuthCacheTxn': ${mosip.certify.access-token-expire-seconds}}
 
 ##-----------------------------VCI related demo configuration---------------------------------------------##
