+++ docker-compose/docker-compose-injistack/config/certify-default.properties
@@ -174,14 +174,14 @@ mosip.certify.common.cache-expire-seconds=3600
 # JWKS cache valid for 5 minutes
 mosip.certify.jwks-cache-expire-seconds=300
 
-mosip.certify.cache.names=userinfo,vcissuance,certificatedatacache,credentialConfig,renderTemplate,jwks,preAuthCodeCache,asMetadataCache
+mosip.certify.cache.names=userinfo,vcissuance,certificatedatacache,credentialConfig,renderTemplate,jwks,preAuthCodeCache,asMetadataCache,preAuthCacheTxn
 # Cache size setup is applicable only for 'simple' cache type.
 # Cache size configuration will not be considered with 'Redis' cache type
 mosip.certify.cache.size={'userinfo': 200, 'vcissuance' : 2000, 'jwks': 1 }
 
 
 # Cache expire in seconds is applicable for both 'simple' and 'Redis' cache type
-mosip.certify.cache.expire-in-seconds={'userinfo': ${mosip.certify.access-token-expire-seconds}, 'vcissuance': ${mosip.certify.access-token-expire-seconds}, 'certificatedatacache': ${mosip.certify.certificatedatacache-expire-seconds}, 'credentialConfig': ${mosip.certify.common.cache-expire-seconds}, 'renderTemplate': ${mosip.certify.common.cache-expire-seconds}, 'jwks': ${mosip.certify.jwks-cache-expire-seconds}}
+mosip.certify.cache.expire-in-seconds={'userinfo': ${mosip.certify.access-token-expire-seconds}, 'vcissuance': ${mosip.certify.access-token-expire-seconds}, 'certificatedatacache': ${mosip.certify.certificatedatacache-expire-seconds}, 'credentialConfig': ${mosip.certify.common.cache-expire-seconds}, 'renderTemplate': ${mosip.certify.common.cache-expire-seconds}, 'jwks': ${mosip.certify.jwks-cache-expire-seconds}, 'preAuthCacheTxn': ${mosip.certify.access-token-expire-seconds}}
 mosip.certify.credential-config.cryptographic-binding-methods-supported={\
   'ldp_vc': {'did:jwk','did:key'},\
   'mso_mdoc': {'cose_key'},\
